refactor: EventManager to FetchCoinsTask
